BackEnd-Snowflake/requirements.txt
fastapi
uvicorn
snowflake-connector-python